Factor startup task logic out into a separate file
